stock/requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai==0.3.1
pytz==2023.3
selenium==4.16.0
webdriver-manager==4.0.1
langchain==0.1.0
langchain_community==0.0.10
yfinance
ta
reportlab>=3.6.12
PyPDF2>=3.0.0
weasyprint>=59.0
pdfkit>=1.0.0
xhtml2pdf>=0.2.10
fpdf2
langchain-google-genai
